services/proposal-composer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
prometheus-client==0.19.0
PyYAML==6.0.1